Refactor workspace logic into `workspace.rs`
