+++ crates/ruff_server/src/lib.rs
@@ -2,8 +2,9 @@
 
 pub use edit::{DocumentKey, NotebookDocument, PositionEncoding, TextDocument};
 use lsp_types::CodeActionKind;
-pub use server::{Server, Workspace, Workspaces};
+pub use server::Server;
 pub use session::{ClientSettings, DocumentQuery, DocumentSnapshot, Session};
+pub use workspace::{Workspace, Workspaces};
 
 #[macro_use]
 mod message;
@@ -16,6 +17,7 @@ mod logging;
 mod resolve;
 mod server;
 mod session;
+mod workspace;
 
 pub(crate) const SERVER_NAME: &str = "ruff";
 pub(crate) const DIAGNOSTIC_NAME: &str = "Ruff";
